requirements.txt
boto3
opensearch-py
azure-storage-blob
sentence-transformers
azure-search-documents==11.4.0
azure-core==1.30.1
openai>=1.30.0
python-dotenv==1.0.1

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langgraph==0.0.15
graphviz>=0.20.1
